integration

+++ JobService/src/main/java/com/JobService/JobService/Model/JobApplicationRequest.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class JobApplicationRequest {
 
-	int jobSeekerId;
+	long jobSeekerId;
 	int jobId;
